Refactored and added readme files
